+++ frontend/src/views/ensemble/barline/BarLineVisualizer.tsx
