+++ frontend/src/views/ensemble/barline/Barline.tsx
